Adds package, setup, cli and scripts
