+++ tests/unit/test_1000g_vcf.py
@@ -1,5 +1,5 @@
 import time
-from tests.unit.conftest import params_from_vcf
+from vrs_anvil import params_from_vcf
 
 
 def test_1000g_vcf(threaded_translator, thousand_genome_vcf, num_threads):
